Moved sheet back into seperate file, now instead has the specific sheet COM object passed to it.
